# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ResumeAI.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Resume/analyze": {
      "post": {
        "tags": [
          "Resume"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeAnalysisRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeAnalysisRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Resume/customize": {
      "post": {
        "tags": [
          "Resume"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "JobDescription": {
                    "type": "string"
                  },
                  "JobDescriptionFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "UseDefaultResume": {
                    "type": "boolean"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "JobDescription": {
                  "style": "form"
                },
                "JobDescriptionFile": {
                  "style": "form"
                },
                "UseDefaultResume": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Resume/optimize": {
      "post": {
        "tags": [
          "Resume"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "UseDefaultResume": {
                    "type": "boolean"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "UseDefaultResume": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResumeAnalysisRequest": {
        "required": [
          "jobDescription",
          "resumeText"
        ],
        "type": "object",
        "properties": {
          "resumeText": {
            "minLength": 1,
            "type": "string"
          },
          "jobDescription": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter 'Bearer {token}' in the field below",
        "scheme": "Bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}